requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.7
httpx==0.25.2
pydantic
python-multipart
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
pypdf
tiktoken